Order of state computation
